### 2022-06-02 #4447 @editorialbot
[PRE REVIEW]: Mat-dp: An open-source Python model for analysing material demand projections and their environmental implications, which result from building low-carbon systems
**Submitting author:** @kcerva (<a href="http://orcid.org/0000-0001-9185-3022">Karla Cervantes Barron</a>)
**Repository:** https://github.com/Mat-dp/mat-dp-core
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
