### 2022-06-07 e56dc96
Updated bibliography details
