+++ paper.bib
@@ -24,5 +24,6 @@
    city = {Freiburg},
    institution = {University of Freiburg},
    month = {1},
-   title = {ODYM_RECC_Model_Docu_V2_2},
+   title = {Documentation of part IVof the RECC model framework: Open Dynamic Material Systems Model for the Resource Efficiency-Climate Change Nexus (ODYM-RECC), v2.2},
+   year = {2020},
 }
