### 2022-06-08 #4447 @crvernon
> @crvernon I also have a question, how can I change the self-generated citation at the end of the pdf? It should say "Cervantes Barron" rather than just "Barron" (this is how last names work for Spanish speakers).
This is what I mean, by the way: "Barron, & Cullen. (2022). Mat-dp: An open-source Python model for analysing material demand projections and their environmental implications, which result from building low-carbon systems. Journal of Open Source Software, 0(0), 4447"

I'll check on this for you.
