### 2022-06-09 #4447 @nick-gorman
> 👋 - @nick-gorman Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @crvernon, yeah happy to be a reviewer!
