### 2022-06-09 #4460 @crvernon
@kcerva regarding your question about name customization:  https://github.com/openjournals/inara/blob/main/docs/names.md

Let me know if you have any questions.
