### 2022-06-13 Mat-dp/mat-dp-core#20 @MoLi7
[JOSS review] Contribution and authorship
Hi @kcerva

I'm reviewing your submission in https://github.com/openjournals/joss-reviews/issues/4460. It seems that the authors of your submission didn't make major contribution to the software directly in your repo, while the major contributors to the software are not listed as co-authors in your paper. This makes me unable to check this one below.

- [ ] Contribution and authorship: Has the submitting author (@kcerva) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Could you provide justification to this? Thanks!
