### 2022-06-13 Mat-dp/mat-dp-core#20 @kcerva
Hi @MoLi7. Thanks for this observation. In this particular repository, my contributions seem small because this repository refers to refactoring of the original Mat-dp model- which was originally in another one of my repositories. This is why the copyright notice gives information about copyright citing the original repository and then talks about modifications. The other contributors to this modified version of the model helped to flesh out attributes of the code that are more in line with proper software development practices and were working under instruction and supervision from myself and Jonathan Cullen. These contributors were externally hired and agreed to cede IP and copyright to the university- a common practice for externally hired software developers. I hope this clarifies things.
