### 2022-06-14 Mat-dp/mat-dp-core#21 @MoLi7
[JOSS review] Community guidelines
There is no guidance for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

I suggest you provide a basic guideline either in Wiki or README.

https://github.com/openjournals/joss-reviews/issues/4460
