### 2022-06-14 Mat-dp/mat-dp-core#22 @MoLi7
[JOSS review] Minor installation issue
In your [installation instruction](https://github.com/Mat-dp/mat-dp-core#installation-and-launch), please clarify whether users need to clone your repo or `pip install` the software in order to launch the model.

https://github.com/openjournals/joss-reviews/issues/4460
