### 2022-06-17 #4460 @nick-gorman
Hi @kcerva, I've just been reviewing the authorship and contributions and from the github pages I can see that @mark-todd and @elliott-huge have made significant contributions to the software. Do you think it would be appropriate for them to be included as co-authors?
