### 2022-06-19 #4460 @nick-gorman
Hi @kcerva, At the start of the paper there is broad claim about the ability of MAT-dp to calculate the  resources needed in any system or supply chain. However, the examples given in the documentation  are all simplistic, demonstrating the functionality on system of real world complexity would makeit much easier to confirm the software's stated functionality claim. The hydro dam example doesn't seem to use main the optimisation functionality, and I'm not sure what it accomplishes that couldn't be done in a simple spreadsheet. Do you think you can add an example that better demonstrates how MAT-dp would be used to solve a real-world problem?
