### 2022-06-19 Mat-dp/mat-dp-core#23 @nick-gorman
[JOSS review] Documentation of process target rate
The documentation is unclear as how setting the rate at which a process consumes resources works. Specifically, when you set an upper and lower bound on the rate I assume there is some penalty to make the solver try and hit the target value, but this isn't stated. Also, if this what's going on, what is the penalty rate being used?

https://github.com/openjournals/joss-reviews/issues/4460
