### 2022-06-19 Mat-dp/mat-dp-core#24 @nick-gorman
[JOSS review] Compiled documentation
Is the documentation compiled into a pdf or html version anywhere? It's a bit hard to follow reading through the docs folder.

https://github.com/openjournals/joss-reviews/issues/4460
