### 2022-06-19 #4460 @nick-gorman
A few extra notes on the paper:
- The summary at the start could be revised for greater clarity:
   - For example, mentioning that the underlying model is a linear program
   - Also the key thing driving the optimisation is the objective function, alongside the constraints you mention
   - I'm not sure it's entirely correct to say that the "system can be optimised by using three types of inbuilt contraints".  The system is optimised by a solver to minimise an objective function subject to constraints.
- I think there is a typo on line 36, do you mean "other" rather than as
- The target audience section seems to continue talking about the functionality of Mat-dp. Could this be re-framed more specifically in terms of potential users?
