### 2022-06-21 b471d5b
Added Contributing section
