### 2022-06-22 168014c
Updated instructions for viewing documentation
