### 2022-06-22 Mat-dp/mat-dp-core#24 @kcerva
Thank you for this observation @nick-gorman. We've added information on the ReadMe on how to see the documentation- either on mkdocs or with a link to a static site.
