### 2022-06-22 9c2bf70
Updated paper based on N Gorman review
Updated line 36, edited the Target Audience Section and edited the description of the model.
