+++ paper.md
@@ -31,17 +31,20 @@ MAT-dp is a python model which aims to calculate the amounts and types of materi
 or resource transformation including those found along any supply chain- but is particularly applied to studying
 the materials needed for building low-carbon systems- and estimate the environmental implications associated to such materials.
 
+Mat-dp contains a linear programming library called mat-dp-core which includes 
+the core classes of each element and their mathematical operations for obtaining results. Mat-dp-core is tailored to the needs of 
+research around material demand for low-carbon systems, helping technically proficient users to explore scenarios and systems 
+relating to their research in python.
+Mat-dp-core has an easy-to-use structure and code base with the mathematical model to let users evaluate and optimise the environmental 
+effects of a given set of resources that are fed into one or more processes.
 
-Mat-dp has an easy-to-use structure and code base with the mathematical model to let users evaluate and optimise the environmental 
-effects of a given set of resources that are fed into one or more processes. This code base is called mat-dp-core, since it contains 
-the core classes of each elements and their mathematical operations for obtaining results.
 
-
-In Mat-dp, the system and its required materials are defined as a series of resources that are fed 
+In mat-dp-core, the system and its required materials are defined as a series of resources that are fed 
 into different processes, which in turn have a defined set of outputs. 
 The outputs can be the result of the process, e.g., energy, materials or greenhouse gas emissions. 
-The system can then be optimised by using three types of constraints built into Mat-dp, depending on the desired 
-output of the process, e.g., a required electricity capacity of an offshore wind installation. 
+The system can then be optimised using an objective function related to a property of the system which will be minimised 
+(e.g., number of runs or cost) subject to three possible types of constraints built into mat-dp-core, 
+depending on the desired output of the process, e.g., a required electricity capacity of an offshore wind installation. 
 The three types of constraints that can be defined are: a given ratio of process runs (Run Ratio Constraint), 
 a given number of resources produced (Resource Constraint), and a given number of process runs (Run Eq Constraint)
 that the system has to comply with.
@@ -59,7 +62,7 @@ Thus, identifying and implementing options for reducing material emissions is re
 
 Mat-dp offers an easy-to-use structure to study material demands, where the types of processes and resources can
 be extended as much as the user needs. To the best of our knowledge, this is the first time that such an extensible
-open-source python model to study materials has been developed. Previous models in the literature and as open-source models have
+open-source python model to study materials has been developed. Previous models in the literature and other open-source models have
 focused on either only a subset of systems (e.g., materials for buildings) or a comprehensive, yet prescribed, 
 set of systems which include some technologies and materials (e.g., ODYM-RECC model [@Pauliuk2020]). The reusable nature of 
 Mat-dp makes it ideal for allowing users to focus on the process(es) they want to investigate, rather than setting up code 
@@ -76,16 +79,23 @@ that are tailored for a specific case-study, which might benefit decision making
 
 # Target Audience
 
+Mat-dp is ideal for academics who work on understanding or reducing environmental effects of different processes,
+government officials who work on national material and emission strategies, or practitioners who work
+on estimating material demands for a given project. Proficiency in python is needed, so the users can
+adapt their requirements in the form needed for the model.
+
 The model facilitates the framing of resource flows and the optimisation of their flows 
 in and out of systems which can then be easily visualised. This allows for users to explore 
 alternative ways to build systems by either reducing or changing their inputs, maximising 
 their outputs or by minimising their emissions or waste.
 
 
-MAT-dp has been used to study the environmental implications of electricity systems, where each 
+Mat-dp has been used to study the environmental implications of electricity systems, where each 
 power plant is considered a process in the model. However, these processes can be anything along 
[…]
