+++ pyproject.toml
@@ -7,8 +7,7 @@ authors = ["Edd Salkield <edd@salkield.uk>"]
 [tool.poetry.dependencies]
 python = ">=3.9,<3.10"
 scipy = "^1.7.1"
-mkdocs = "^1.2.3"
-mkdocs-material = "^8.1.7"
+numpy = "^1.23.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.0"
@@ -19,6 +18,8 @@ black = "^21.9b0"
 isort = "^5.9.3"
 pre-commit = "^2.15.0"
 autoflake = "^1.4"
+mkdocs = "^1.2.3"
+mkdocs-material = "^8.1.7"
 
 [tool.black]
 line-length = 79
