### 2022-07-14 6086367
Updated toml with info needed for PyPi
