### 2022-07-15 b81a208
Update README.md
Changed instructions to use pip or for developers using poetry
