### 2022-07-15 5d66fc6
Updated ReadMe with installation instructions and prepared repo for PyPi
