+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "mat-dp-core"
-version = "0.1.0"
+version = "0.1.1"
 description = ""
 license = "MIT"
 authors = ["Edd Salkield <edd@salkield.uk>"]
