### 2022-07-15 Mat-dp/mat-dp-core#22 @kcerva
Hi @MoLi7 we have now added mat-dp-core to PyPi and have therefore modified the instructions to say that the user can install the library using pip.
