### 2022-07-15 Mat-dp/mat-dp-core#23 @kcerva
Hi @nick-gorman. I think this is a misunderstanding of how the model works. The rates at which processes consume resources are defined by the user (e.g., as kg materials/MW). Upper and lower bounds of these rates (or material intensities) can be given too. There is no "penalty" per se defined in the model (i.e., this is not a linear optimisation defined with a penalty function). The way it works, is just the usual way where an objective function is defined. The function represents the target of the solver, but the values of the process runs can vary anywhere within the specified upper and lower bounds to minimise/maximise the target function. The exact value specified in this case is just the result if no bounds were set. Please see [here](https://client.dreamingspires.dev/mat_dp_core/Function%20Reference/objective/) for documentation on the objective function.

We have also included error messages for when problems are overconstrained or unbounded so the user can make the required changes. Please see more [here](https://client.dreamingspires.dev/mat_dp_core/Function%20Reference/exceptions/).

Hope this answers your question.
