### 2022-07-15 2713ad0
Edited ReadMe to show statement of need and fixed minor stylistic issues of text
