+++ docs/index.md
@@ -1,38 +1,39 @@
-# MATerial Demand Projections Core
+# Material Demand Projections Core
 
 
 ![](res/mat%20dp%20core%20banner.png)
 
-Welcome to the MAT-DP Core. This repository sits at the centre of a wider project to deliver user-friendly and open-access software solutions to environment and energy research.  
-While this software *could* be used for many applications, MAT-DP Core and its documentation has been designed with the research of *'environmental impact and material usage in the building of low-carbon systems'* in mind.
+Welcome to the Mat-dp core. This repository sits at the centre of a wider model called Mat-dp (Material Demand Projections) which is part of a project to deliver user-friendly and open-access software solutions to environment and energy research.  
+While this software *could* be used for many applications, Mat-dp core and its documentation have been designed to research *'environmental impacts and material use in the building of low-carbon systems'*.
 
-## **What is MAT-DP Core?**
+## **What gap does Mat-dp fill?**
 
-MAT-DP Core is a linear programming library - available at [this Github repository](https://github.com/dreamingspires/mat-dp-core) - tailored to the needs of research around material demand for low-carbon systems.  
-MAT-DP Core helps technically proficient users to explore scenarios and systems relating to their research in Python.
+Materials are essential for creating the systems used in everyday life, yet, manufacturing materials emits important quantities of greenhouse gases (GHG). Emissions from material production have been estimated to increase from 5 gigatons (Gt) of CO2-equivalent to 11 Gt between 1995 and 2015. In turn, global emissions due to material production rose from 15 per cent to 23 per cent in the same period [2], while materials are estimated to contribute to over half of GHG emissions from industry[1].
+Thus, identifying and implementing options for reducing material emissions are required.
+
+Mat-dp offers an easy-to-use structure to study material demands, where the types of processes and resources can be extended as much as the user needs. To the best of our knowledge, this is the first time that such an extensible open-source python model to study materials has been developed. Previous models in the literature and other open-source models have focused on either only a subset of systems (e.g., materials for buildings) or a comprehensive, yet prescribed, set of systems which include some technologies and materials (e.g., ODYM-RECC model [3]). The reusable nature of Mat-dp makes it ideal for allowing users to focus on the process(es) they want to investigate, rather than setting up code or a mathematical model from scratch. The benefits of such reusability to advancing research in this scientific discipline cannot be understated.
+
+Mat-dp is capable of adapting the temporal and regional specifications of users by creating the processes according to prescribed names and using indices for the desired categories, once the Mat-dp library has been added to a working example. Mat-dp also allows for different material mixes, technology changes or improvements over time, and recycled material content in a given process to be included in the model. These capabilities allow for models to be built that are tailored for a specific case-study, which might benefit decision making.
+
+## **What is Mat-dp-core?**
+
+Mat-dp contains a linear programming library called Mat-dp core which includes the core classes of each element and their mathematical operations for obtaining results. Mat-dp-core has an easy-to-use structure and code base with the mathematical model to let users evaluate and optimise the environmental effects of a given set of resources that are fed into one or more processes.
+
[…]
