+++ README.md
@@ -1,7 +1,7 @@
 # Mat-dp core
 ## *Material Demand Projections Model*
 
-Welcome to the Mat-dp core. This repo represents the core of the MAT-dp project, which aims to deliver user-friendly and open-access software to study the environmental implications of materials used for building low-carbon systems. 
+Welcome to the Mat-dp core. This repo represents the core of the Mat-dp project, which aims to deliver user-friendly and open-access software to study the environmental implications of materials used for building low-carbon systems. 
 
 # Installation and launch
 
