### 2022-07-18 Mat-dp/mat-dp-core#23 @nick-gorman
Does this mean that the target value is redundant if an upper and lower bound are provided? If so, maybe the docs should state this?
