### 2022-07-19 #4460 @nick-gorman
Hi @kcerva and @crvernon,

I think the only outstanding point is a question I still have on the documentation: https://github.com/Mat-dp/mat-dp-core/issues/23

Regarding example complexity happy to accept it as is, they are definitely sufficient to explain how to use Mat-dp.

However, just as general feedback, I still think it would be really useful to see Mat-dp simulating a system of real world complexity, or answering a real-world question. From the ones you've provided it's hard to see the use case. Generally, models of energy systems capture much more of the system complexity than the one presented in the Function Reference, Calliope is great example of a tool that does this, and lets the user model multiple resource flows and co2 production: https://www.callio.pe/model-gallery/.  The complex example in the cookbook leaves me a bit confused, the hydro damn is modelled as a process that consumes raw materials and produces a fixed quantity of energy. However, the role of hydro in energy systems is much more nuanced than this, for instances it will have huge value in low carbon systems providing energy when wind and solar production are low. If modelling doesn't capture these dynamics it will provide misleading answers about how best to build low carbon energy systems.

Pardon the rant. I'm really not trying to question the value of Mat-dp. Just trying to explain why different examples might better communicate how Mat-dp could be used.

Cheers,
Nick
