### 2022-07-19 Mat-dp/mat-dp-core#23 @kcerva
No, the target value is used in the optimisation for sure, and the upper and lower bound are optional and can be used to find some sort of uncertainty range.
