### 2022-07-19 #4460 @kcerva
Thanks @nick-gorman! Just answered the other question.

Also, I think we should stress that this isn't just another energy systems model, it is tailored towards identifying the materials needed for building systems. Therefore, the variability that you mention associated to intermittency of renewables is not an issue, since power plants are designed for a given capacity or power demand (usually taking the capacity and some operational factors into account)- that is why our hydro example has power demand, but of course capacity can be used too. Thus, even if the operational aspect of a power plant involves changes for energy modelling, the materials used to initially build the power plants remain unchanged. Hope this makes sense. This should also help explain more why our model brings an extra useful tool for integrated energy planning.

Thanks for pointing us to calliope, though. It looks great!
