### 2022-07-19 Mat-dp/mat-dp-core#23 @nick-gorman
Do you think the meaning of these inputs and what they do needs to be better documentation? Honestly, I'm still unsure.

If the target values are still used in the optimization, even when the upper and lower bound are provided, how does that work considering what you said before "the values of the process runs can vary anywhere within the specified upper and lower bounds to minimise/maximise the target function". How is the target value being used?
