### 2022-07-21 Mat-dp/mat-dp-core#23 @nick-gorman
ok, now I think I understand it "their use involves solving the optimisation at the specified bounds", that make sense. Is this documented somewhere?
