### 2022-07-28 Mat-dp/mat-dp-core#23 @kcerva
Yes, there is documentation in the [Processes section](https://client.dreamingspires.dev/mat_dp_core/Function%20Reference/processes/). See the example code in particular, which is even more explicit. Then, the [Measure section](https://client.dreamingspires.dev/mat_dp_core/Function%20Reference/measure/) complements this by giving options of when/how to "turn on/off" the bounds.
