### 2022-07-28 Mat-dp/mat-dp-core#23 @nick-gorman
OK, I guess I was just trying to communicate that for me as a new user those pieces of documentation didn't really communicate what as going on clearly. All good.
