### 2022-08-09 0e08a1a
Edited print statement instructions of ReadMe
