### 2022-08-09 #4460 @kcerva
Sorry for the slight delay in replying, but thanks @nick-gorman for all your comments!

@crvernon thanks for your points! I'm slowly addressing them and will let you know once I'm done, but an immediate comment that comes to mind about the python vs Python part is that the pre-reviewer wanted the opposite. Please see the 2nd/June comment by Daniel on it [here](https://github.com/openjournals/joss-reviews/issues/4447#issuecomment-1145090076). So, which one is the right one?
