+++ paper.bib
@@ -21,6 +21,7 @@
 }
 @report{Pauliuk2020,
    author = {Stefan Pauliuk},
+   doi = {10.31235/osf.io/y4xcv},
    city = {Freiburg},
    institution = {University of Freiburg},
    month = {1},
