### 2022-08-16 #4460 @kcerva
Thank you @crvernon! I think I've now addressed all your points. Please see below:

- "python" should be have a capital "P" when discussed in the paper -> As agreed.
- See the missing DOI above and confirm that the suggested one 10.31235/osf.io/y4xcv is or is not a viable option -> Added suggested DOI to bib file. Will regenerate pdf.
- Place the link to your documentation in the "About" section of your GitHub repo along with a description of your repository. This helps folks easily find your doc site -> Done
- Was there a reason that you didn't set up continuous integration testing via GitHub actions or the like for your package? Just curious. This would help keep things stable on your main branch moving forward. Let me know if you would like me to send some examples of how to do this. -> As far as I understand, pytest is easy to implement and it's good for a small set of unit tests (like our current package). I don't know enough about continuous integration testing, so more information is definitely welcome! Once I have a look, perhaps that can be incorporated in the next version of mat-dp-core.
- Step 5 in the step-by-step code in your README produces the following error. -> The error was due to not having updated that particular instruction after changing the code, but I've now updated it and even added an extra suggestion on how to use it, so it should run now.
- More robust docstrings for your classes, methods, functions that will bring clarity for future users. -> Thank you for this, I will look into it for the next update.
