### 2022-08-16 #4460 @crvernon
@kcerva - we are almost there! Next is just setting up the archive for your new release.

**So here is what we have left to do:**

- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ ] Please list the DOI of the archived version here
