### 2022-08-17 #4460 @kcerva
Thanks @crvernon!
I've done the release and used Zenodo. The DOI is the following: 10.5281/zenodo.7002733

I also have a question. Hired developers were involved in modifying the software to its current version (who ceded their IP and copyright to the University of Cambridge). I've then reflected this in the Copyright part of the repo (which says that the original copyright was mine and Jonathan Cullen's and the modifications are copyrighted to the University. I acknowledged the external developers there as having made modifications). For the paper, Jonathan and myself were involved in writing it. So, for Zenodo I've kept those same names, but I just want to double-check this is correct.
For more context, please see [here](https://github.com/Mat-dp/mat-dp-core/issues/20).
