### 2022-08-17 #4460 @crvernon
Thanks @kcerva .  I lean towards the inclusion of all who have contributed to the software in its current form unless they have stated that they do not wish to be included.  If you believe the ceding of the IP is indicative of their wishes to not be included as co-authors, then we may proceed in the current format.  If not, please add in the additional authors.  Let me know when you have decided and we will move forward.  Thanks!
