### 2022-08-20 #4460 @arfon
Apologies for the back-and-forth here folks but house style (and accepted convention as far as I know) is _Python_ for the language name, `python` when talking about the command-line interaction.

As such, please merge this PR @kcerva: https://github.com/Mat-dp/mat-dp-core/pull/25
