### 2022-11-21 #4460 @kcerva
Hello @arfon. I am doing some admin of publications and I realised that the ORCID associated to this paper for my name is wrong. It should be https://orcid.org/my-orcid?orcid=0000-0001-9185-3022, but instead it has someone else's account. Is there a way to fix this?
