### 2022-11-30 #4460 @xuanxu
@kthyng the orcid should be updated in the metadata section of the `paper.md` file. Then reaccepting the submission (`@editorialbot reaccept`) will regenerate the paper with the correct orcid.
